Fix `load_maze()`
If there is no start or goal cell implied by the maze file, the method
does not initiate the start and goal attributes. This commit fixes this
issue by initializing the start and goal attributes to a random cell if
this happens.

Also, the method has more issues if the goal attribute has been
initialized beforehand. The new goal cell(s), instead of being set as
the only goal cell(s), are simply added to the previously initialized
goal cell(s). This commit also fixes this issue.